Extract HiddenAvailability
